requirement.txt
alabaster==0.7.12
aniso8601==7.0.0
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
Babel==2.8.0
certifi==2019.11.28
chardet==3.0.4
coverage==5.0.3
Django==3.0.2
django-filter==2.2.0
django-graphql-jwt==0.3.0
docutils==0.16
Faker==0.9.1
graphene==2.1.8
graphene-django==2.8.1
graphql-core==2.3.1
graphql-relay==2.0.1
idna==2.8
imagesize==1.2.0
isort==4.3.21
Jinja2==2.11.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mixer==6.1.3
more-itertools==8.2.0
packaging==20.1
pluggy==0.13.1
promise==2.3
psycopg2==2.8.4
py==1.8.1
Pygments==2.5.2
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.6
pytest==5.3.5
pytest-cov==2.8.1
pytest-django==3.8.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
snowballstemmer==2.0.0
Sphinx==2.3.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
text-unidecode==1.2
urllib3==1.25.8
wcwidth==0.1.8
wrapt==1.11.2
